Plots the dependency graph

The dependency structure depends on feature which determines which repositories are checked out. It starts from the project and discovers all reachable upstream and downstream dependencies.

dependency_graph(
  project = ".",
  feature = NULL,
  local_repos = get_local_pkgs_from_config(),
  verbose = 0
)

Arguments

project
(character) directory of project (for which to restore the dependencies according to feature); must be a git repository.

feature
(character) feature we want to build; inferred from the branch of the project if not provided; warning if not consistent with current branch of project

local_repos
(data.frame) repositories that should be taken from local rather than cloned; columns are repo, host, directory

verbose
(numeric) verbosity level, incremental; (0: None, 1: packages that get installed + high-level git operations, 2: includes git checkout infos)

Value

depending on return_table_only (see above); either data.frame or a list with the following elements:

df: data.frame with columns repo, host, type, branch, where type is:

current (project),

upstream (of project),

downstream (of project),

other (the remaining, e.g. downstream dependencies of upstream dependencies)

and branch is the branch corresponding to feature and becomes local (branch) if package is in local_repos.

graph: visNetwork graph with arrows going in the direction from downstream to upstream. Arrows are solid if both upstream and downstream list each other; they are dashed if either does not list the other:

if upstream does not list downstream, blue dashed

if downstream does not list upstream, red dashed (this means that the downstream repo's yaml needs to be fixed) Tooltips give more information about nodes and edges.

deps: deps: upstream and downstream dependencies of each node

Examples

if (FALSE) {
dependency_graph()
}