Gadget or Shiny app to select the dependencies to install

The dependencies are obtained by traversing the upstream and downstream repositories in the package's staged dependencies yaml files starting from project.

install_deps_app(
  default_repo = NULL,
  default_host = "https://github.com",
  default_ref = "main",
  fallback_branch = "main",
  run_gadget = TRUE,
  run_as_job = TRUE,
  verbose = 1,
  install_external_deps = TRUE,
  upgrade = "never",
  ...
)

Arguments

default_repo
(character) the repository name for the dependency graph to be created for, for example, "openpharma/stageddeps.water". If NULL this must be entered by app user and can always be changed by the user.

default_host
(character) the host for the repository for the dependency graph to be created for by default "https://github.com". If NULL this must be entered by app user and can always be changed by the user.

default_ref
(character) default ref (branch/tag), see also the parameter ref of \link{dependency_table}. If NULL this must be entered by app user and can always be changed by the user.

fallback_branch
(character) the default branch to try to use if no other matches found

run_gadget
(logical) whether to run the app as a gadget

run_as_job
(logical) whether to run the installation as an RStudio job.

verbose
verbosity level, incremental; from 0 (none) to 2 (high)

install_external_deps
logical to describe whether to install external dependencies of packages using remotes::install_deps.

upgrade
argument passed to remotes::install_deps, defaults to 'never'.

...
Additional args passed to `remotes::install_deps.

Value

shiny.app or value returned by app (executed as a gadget)

Examples

if (FALSE) {
  install_deps_app("openpharma/stageddeps.food")
}