Install dependencies job

Install dependencies job

install_deps_job(
  project = ".",
  project_type = "local",
  verbose = 1,
  create_args = list(renv_profile = Sys.getenv("RENV_PROFILE")),
  ...
)

Arguments

project

(character) If project_type is local then directory of project (for which to calculate the dependency structure); must be a git repository. If project_type is repo@host then should be character of the form openpharma/stageddeps.food@https://github.com If host is not included in the string then the default https://github.com is assumed.

project_type

(character) See project argument

verbose

(numeric) verbosity level, incremental; (0: None, 1: packages that get installed + high-level git operations, 2: includes git checkout infos)

create_args

named list - additional arguments passed to dependency_table function

...

Arguments passed on to install_deps

dep_structure

(dependency_structure) output of function dependency_table; uses dep_structure$table to infer the packages to apply action to and infer installation order; uses dep_structure$deps to infer upstream dependencies

install_project

(logical) whether to also install the current package (i.e. the package named in dependency_structure$current_pkg), ignored unless install_direction = "upstream" (because downstream deps automatically install all their upstream deps)

install_direction

"upstream", "downstream" or "all"; which packages to install (according to dependency structure). By default this is only "upstream"

install_external_deps

logical to describe whether to install external dependencies of package using remotes::install_deps() (or renv::install() if inside an renv environment) .

upgrade

argument passed to remotes::install_deps(), defaults to 'never'. Ignored if inside an renv environment.

package_list

(character) If not NULL, an additional filter, only packages on this list will be considered and their dependencies installed if needed (advanced usage only).

dry

(logical) dry run that outputs what would happen without actually doing it.

See also

Examples

if (FALSE) {
install_deps_job()
install_deps_job(create_args = list(ref = "6_makegraph@main"))

# install all dependencies
install_deps_job(create_args = list(direction = "all"))
install_deps_job(dry_install = TRUE)
}