Run an action on the packages

Run an action on the packages

run_package_actions(
  pkg_actions,
  internal_pkg_deps,
  dry = FALSE,
  install_external_deps = TRUE,
  upgrade = "never",
  rcmd_args = NULL,
  artifact_dir = NULL,
  verbose = 0,
  ...
)

Arguments

pkg_actions

sorted data.frame with columns cache_dir pointing to package source, sha, installable and actions, processed from top to bottom in order Each actions is a subset of "test", "build", "check", "install";

if test: run with devtools::test

if build: R CMD build

if check: rcmdcheck with output to command line (if no build action), otherwise R CMD check on tarball

if install: R install based on remotes::install (if no build action), otherwise R CMD INSTALL on tarball

internal_pkg_deps

packages to not install (when install_external_deps = TRUE)

dry

logical, if FALSE (default) run the actions, if TRUE do not

install_external_deps

logical to describe whether to install external dependencies of package using remotes::install_deps() (or renv::install() if inside an renv environment) .

upgrade

argument passed to remotes::install_deps(), defaults to 'never'. Ignored if inside an renv environment.

rcmd_args

list with names build, check, install which are vectors that are passed as separate arguments to the R CMD commands

artifact_dir

directory to store log files, only when actions include build; action test only outputs to the console

verbose

verbosity level, incremental - from 0 (none) to 2 (high)

...

Additional args passed to remotes::install_deps() Ignored if inside an renv environment.